backend/requirements.txt
Django
djangorestframework
python-decouple
celery
redis
djangorestframework-simplejwt
Pillow
django-redis
pytest
pytest-django